+++ cloud-service/src/main/java/com/example/cloudservice/domain/AuthOperationLog.java
@@ -43,7 +43,7 @@ public class AuthOperationLog implements Serializable {
     /**
      * 是否执行成功(0失败1成功)
      */
-    private Byte succeed;
+    private Short succeed;
 
     /**
      * 具体消息备注
